streamlit-docker/requirements.txt
altair==4.2.2
amazon-textract-caller==0.0.28
amazon-textract-prettyprinter==0.1.2
amazon-textract-response-parser==0.1.45
attrs==23.1.0
blinker==1.6.2
boto3==1.26.121
botocore==1.29.121
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
mdurl==0.1.2
numpy==1.24.3
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
rich==13.3.5
s3transfer==0.6.0
six==1.16.0
smmap==5.0.0
streamlit==1.21.0
tabulate==0.9.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
zipp==3.15.0
